Version 0.0.4.12:
- GCC-4.7.2-Kompatible Userlib
- Fehler durch ungenaue Fließkommarechnungen bei clang behoben
- Compilerfehler mit _FAT_DIAGNOSIS_ behoben
- disk_t::size als 64-bit Integer - Festplatten können größer als 4 GiB sein. %Sa, %Sg, %Sm und %Sk erfordern 64-bit int.
- Verschiebung der PCI-Vendorliste in VS10-Projektmappe nachgetragen
- VCcompatibility.h ergänzt und aktualisiert (GCC 4.7.2 als Standard)

git-svn-id: svn://svn.code.sf.net/p/prettyos/code/trunk/Source@1397 31957799-4632-4615-9793-72db6d48bcac

+++ kernel/storage/devicemanager.h
@@ -30,7 +30,7 @@ typedef struct disk
 {
     diskType_t*  type;
     partition_t* partition[PARTITIONARRAYSIZE]; // 0 if partition is not used
-    size_t       size;                          // size of disk in bytes
+    uint64_t     size;                          // size of disk in bytes
     char         name[15];
     void*        data;                          // Contains additional information depending on disk-type
     uint32_t     accessRemaining;               // Used to control motor
